moved all profile related routes to route folder in profileInfo.js, -currently work in progress, going to separate this into smaller sections, doing this incrementally
