+++ app_js_routes.txt
@@ -1,0 +1,49 @@
+/create-profile
+
+/dashboard
+
+/dashboard/ads
+
+/dashboard/ads/enable
+
+/dashboard/ads/create
+
+/dashboard/ads/delete
+
+/dashboard/ads/edit
+
+/dashboard/ads/sort  <-- change adSortOrder to this
+
+/
+
+/login
+
+/register
+
+/logout
+
+/profile
+
+/profile/header
+
+/profile/about
+
+/profile/website
+
+/profile/instruments
+
+/profile/instrument/update <-- make instruments plural
+
+/profile/instrument/add <-- make instruments plural
+
+/profile/worksamples
+
+/profile/worksamples/music
+
+/profile/worksamples/video
+
+/profile/levels
+
+/profile/basic
+
+/profile/basic/create
